+++ src/options.ts
@@ -1,25 +1,46 @@
 import type { Consola } from 'consola'
 import { defu } from 'defu'
+import { relative } from 'pathe'
+import { Integrations as ServerIntegrations } from '@sentry/node'
+import * as PluggableIntegrations from '@sentry/integrations'
 import type { Options } from '@sentry/types'
 import type { Nuxt } from '@nuxt/schema'
 import { resolveAlias } from '@nuxt/kit'
-import * as Integrations from '@sentry/integrations'
 import type { NuxtOptions } from '@nuxt/types'
-import type { IntegrationsConfiguration, LazyConfiguration, TracingConfiguration } from '../../types/sentry'
-import type { ModuleConfiguration } from '../../types'
+import type { Integrations, LazyConfiguration, TracingConfiguration } from '../types/sentry'
+import type { ModuleConfiguration } from '../types'
 import { canInitialize } from './utils'
 
-export const PLUGGABLE_INTEGRATIONS = ['CaptureConsole', 'Debug', 'Dedupe', 'ExtraErrorData', 'ReportingObserver', 'RewriteFrames']
-export const BROWSER_INTEGRATIONS = ['InboundFilters', 'FunctionToString', 'TryCatch', 'Breadcrumbs', 'GlobalHandlers', 'LinkedErrors', 'HttpContext']
-const SERVER_INTEGRATIONS = ['CaptureConsole', 'Debug', 'Dedupe', 'ExtraErrorData', 'RewriteFrames', 'Modules', 'Transaction']
+// Enabled by default in Vue - https://docs.sentry.io/platforms/javascript/guides/vue/configuration/integrations/default/
+export const BROWSER_INTEGRATIONS = ['Breadcrumbs', 'Dedupe', 'FunctionToString', 'GlobalHandlers', 'HttpContext', 'InboundFilters', 'LinkedErrors', 'TryCatch']
+// Optional in Vue - https://docs.sentry.io/platforms/javascript/guides/vue/configuration/integrations/plugin/
+export const BROWSER_PLUGGABLE_INTEGRATIONS: Array<keyof typeof PluggableIntegrations> = ['CaptureConsole', 'Debug', 'ExtraErrorData', 'HttpClient', 'ReportingObserver', 'RewriteFrames']
+// Enabled by default in Node.js - https://docs.sentry.io/platforms/node/configuration/integrations/default-integrations/
+const SERVER_INTEGRATIONS: Array<keyof typeof ServerIntegrations> = ['Console', 'ContextLines', 'FunctionToString', 'Http', 'InboundFilters', 'LinkedErrors', 'LocalVariables', 'Modules', 'OnUncaughtException', 'OnUnhandledRejection', 'RequestData']
+// Optional in Node.js - https://docs.sentry.io/platforms/node/configuration/integrations/pluggable-integrations/
+const SERVER_PLUGGABLE_INTEGRATIONS: Array<keyof typeof PluggableIntegrations> = ['CaptureConsole', 'Debug', 'Dedupe', 'ExtraErrorData', 'RewriteFrames', 'Transaction']
 
-const filterDisabledIntegrations = (integrations: IntegrationsConfiguration): string[] => Object.keys(integrations).filter(key => integrations[key])
+function filterDisabledIntegrations<T extends Integrations> (integrations: T): Array<keyof T> {
+  return getIntegrationsKeys(integrations).filter(key => integrations[key])
+}
+
+function getIntegrationsKeys<T extends Integrations> (integrations: T): Array<keyof T> {
+  return Object.keys(integrations)
+}
 
 async function getApiMethods (packageName: string): Promise<string[]> {
   const packageApi = await import(packageName)
 
   const apiMethods: string[] = []
   for (const key in packageApi) {
+    if (key === 'default') {
+      for (const subKey in packageApi[key]) {
+        if (typeof packageApi[key][subKey] === 'function') {
+          apiMethods.push(subKey)
+        }
+      }
+      continue
+    }
     if (typeof packageApi[key] === 'function') {
       apiMethods.push(key)
     }
@@ -96,11 +117,13 @@ function resolveTracingOptions (options: ModuleConfiguration, config: NonNullabl
   }
 
   options.tracing = tracingOptions
+  // Enable tracing for `Http` integration.
+  options.serverIntegrations = defu(options.serverIntegrations, { Http: { tracing: true } })
 }
 
-export type resolvedClientOptions = {
-  PLUGGABLE_INTEGRATIONS: string[]
+export type ResolvedClientOptions = {
   BROWSER_INTEGRATIONS: string[]
+  BROWSER_PLUGGABLE_INTEGRATIONS: string[]
   dev: boolean
   runtimeConfigKey: string
   config: Options
@@ -115,23 +138,24 @@ export type resolvedClientOptions = {
   integrations: Record<string, unknown>
 }
 
-export async function resolveClientOptions (nuxt: Nuxt, moduleOptions: ModuleConfiguration, logger: Consola): Promise<resolvedClientOptions> {
+export async function resolveClientOptions (nuxt: Nuxt, moduleOptions: ModuleConfiguration, logger: Consola): Promise<ResolvedClientOptions> {
   const options = moduleOptions
-  options.config = defu(options.clientConfig, options.config)
+  let config = defu({}, options.config)
 
   let clientConfigPath: string | undefined
   if (typeof (options.clientConfig) === 'string') {
     clientConfigPath = resolveAlias(options.clientConfig)
+    clientConfigPath = relative(nuxt.options.buildDir, clientConfigPath)
   } else {
-    options.config = defu(options.clientConfig, options.config)
+    config = defu(options.clientConfig, options.config)
   }
 
   const apiMethods = await getApiMethods('@sentry/vue')
   resolveLazyOptions(options, apiMethods, logger)
-  resolveTracingOptions(options, options.config)
+  resolveTracingOptions(options, config)
 
-  for (const name of Object.keys(options.clientIntegrations)) {
-    if (!PLUGGABLE_INTEGRATIONS.includes(name) && !BROWSER_INTEGRATIONS.includes(name)) {
+  for (const name of getIntegrationsKeys(options.clientIntegrations)) {
+    if (![...BROWSER_INTEGRATIONS, ...BROWSER_PLUGGABLE_INTEGRATIONS].includes(name)) {
       logger.warn(`Sentry clientIntegration "${name}" is not recognized and will be ignored.`)
       delete options.clientIntegrations[name]
     }
@@ -141,19 +165,20 @@ export async function resolveClientOptions (nuxt: Nuxt, moduleOptions: ModuleCon
   if (options.customClientIntegrations) {
     if (typeof (options.customClientIntegrations) === 'string') {
       customClientIntegrations = resolveAlias(options.customClientIntegrations)
+      customClientIntegrations = relative(nuxt.options.buildDir, customClientIntegrations)
     } else {
       logger.warn(`Invalid customClientIntegrations option. Expected a file path, got "${typeof (options.customClientIntegrations)}".`)
     }
   }
 
   return {
-    PLUGGABLE_INTEGRATIONS,
     BROWSER_INTEGRATIONS,
+    BROWSER_PLUGGABLE_INTEGRATIONS,
     dev: nuxt.options.dev,
     runtimeConfigKey: options.runtimeConfigKey,
     config: {
       dsn: options.dsn,
-      ...options.config,
+      ...config,
     },
     clientConfigPath,
     lazy: options.lazy,
@@ -170,18 +195,18 @@ export async function resolveClientOptions (nuxt: Nuxt, moduleOptions: ModuleCon
   }
 }
 
-export type resolvedServerOptions = {
+export type ResolvedServerOptions = {
   config: Options
   apiMethods: string[]
   lazy: boolean | LazyConfiguration
   logMockCalls: boolean
 }
 
-export async function resolveServerOptions (nuxt: Nuxt, moduleOptions: ModuleConfiguration, logger: Consola): Promise<resolvedServerOptions> {
+export async function resolveServerOptions (nuxt: Nuxt, moduleOptions: ModuleConfiguration, logger: Consola): Promise<ResolvedServerOptions> {
   const options = moduleOptions
 
-  for (const name of Object.keys(options.serverIntegrations)) {
-    if (!SERVER_INTEGRATIONS.includes(name)) {
+  for (const name of getIntegrationsKeys(options.serverIntegrations)) {
+    if (![...SERVER_INTEGRATIONS, ...SERVER_PLUGGABLE_INTEGRATIONS].includes(name)) {
       logger.warn(`Sentry serverIntegration "${name}" is not recognized and will be ignored.`)
       delete options.serverIntegrations[name]
     }
@@ -202,13 +227,21 @@ export async function resolveServerOptions (nuxt: Nuxt, moduleOptions: ModuleCon
 
   const defaultConfig = {
     dsn: options.dsn,
-    intergrations: [
+    integrations: [
       ...filterDisabledIntegrations(options.serverIntegrations)
         .map((name) => {
           const opt = options.serverIntegrations[name]
-          // TODO Fix this type
-          // eslint-disable-next-line import/namespace
-          return Object.keys(opt as Record<string, unknown>).length ? new Integrations[name](opt) : new Integrations[name]()
+          try {
+            if (SERVER_INTEGRATIONS.includes(name)) {
+              return Object.keys(opt as Record<string, unknown>).length ? new ServerIntegrations[name](opt) : new ServerIntegrations[name]()
+            } else {
+              // TODO Fix this type
+              // eslint-disable-next-line import/namespace
+              return Object.keys(opt as Record<string, unknown>).length ? new PluggableIntegrations[name](opt) : new PluggableIntegrations[name]()
+            }
+          } catch (error) {
+            throw new Error(`Failed initializing server integration "${name}".\n${error}`)
+          }
         }),
       ...customIntegrations,
     ],
@@ -224,14 +257,14 @@ export async function resolveServerOptions (nuxt: Nuxt, moduleOptions: ModuleCon
     }
   }
 
-  options.config = defu(defaultConfig, options.config, options.serverConfig, getRuntimeConfig(nuxt, options))
+  const config = defu(defaultConfig, options.config, options.serverConfig, getRuntimeConfig(nuxt, options))
 
   const apiMethods = await getApiMethods('@sentry/node')
   resolveLazyOptions(options, apiMethods, logger)
   resolveTracingOptions(options, options.config)
 
   return {
-    config: options.config,
+    config,
     apiMethods,
     lazy: options.lazy,
     logMockCalls: options.logMockCalls, // for mocked only